feat: #6 Adds home detail
