--- src/components/BlogSummary.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-function BlogSummary(): React.ReactElement<{}> {
-  return <div>BlogSummary</div>
-}
-
-export default BlogSummary
