+++ src/components/BlogSummary/index.css
@@ -1,0 +1,6 @@
+.summary-container {
+    display: grid;
+    grid-template-columns: 50% 50%;
+    grid-gap: 10px;
+    padding: 3rem;
+}
