+++ src/index.css
@@ -6,7 +6,7 @@
 
   color-scheme: dark;
   color: #000;
-  background-color: #ddd;
+  background-color: #e6ebe0;
 
   font-synthesis: none;
   text-rendering: optimizeLegibility;
